getUserId update
When determining an event's user through which sensor the event comes
from, we will now assume it is the currently attended user. If we do not
have an attended user, we will go through the regular route of picking
the user closest  to the sensor's location in the situation model.